+++ changelog.md
@@ -1,4 +1,11 @@
 
+# V 0.3.0
+
+- [x] Added support for VCSC index types
+- [x] Added support for random access reads
+- [x] Added support for reading/writing I/VCSC files
+- [x] Started work on documation and comments
+
 # V 0.2.1
 
 - [x] Added min, max, and sum to VCSC and IVCSC